requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.1
openpyxl==3.1.2
WeasyPrint==60.1
matplotlib==3.8.0
libsql-client>=0.6.0
Jinja2==3.1.2
